Fixed cursor cliping after application initialization.
It was possible that the application window was the active window but not the foreground window.
In this strange situations, the cursor was clipped to application area which interferes with other application.
This happens when starting the game from Visual Studio and switching focus (alt+tab) to another window before the game window becomes visible.